# api/openapi.yaml
openapi: 3.1.0
info:
  title: Plataforma MOOC API
  description: >
    Official REST API specification for the Massive Open Online Course (MOOC) platform (ISIS4426 Cloud Course).
    All endpoints require HTTP/JSON communications and strictly reside under the `/api/v1` path prefix.
  version: 1.0.0
  contact:
    name: MOOC Engineering Team
    email: support@mooc-platform.example.com
    url: https://mooc-platform.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: Primary API v1 Endpoint Base Path
...
paths:
...
  /admin/users:
    get:
      tags:
        - Users
      summary: List accounts
      description: >
        Administrative listing of accounts, newest first, with optional filters
        by role and status.


        Pagination is by cursor rather than by offset: an offset shifts whenever
        a row is inserted or removed, which makes a client silently skip or
        repeat users while it pages through them.
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          required: false
          schema:
            type: string
            enum: [administrador, profesor, estudiante]
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending_verification, active, suspended]
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: One page of accounts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '500':
          $ref: '#/components/responses/500InternalServerError'
  /admin/users/{user_id}:
    get:
      tags:
        - Users
      summary: Read a single account
      description: >
        Returns the account together with its `ETag`, which is what a client
        needs before attempting a conditional write.


        A request carrying `If-None-Match` with the current tag answers `304`
        without a body, so polling for changes costs almost nothing.
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: The account.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
        '304':
          $ref: '#/components/responses/304NotModified'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404NotFound'
        '500':
          $ref: '#/components/responses/500InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Opaque session token issued by `POST /auth/login`. It is not a JWT: the
        session is looked up on every request, which is what makes revocation
        take effect immediately.
  parameters:
    CursorParam:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Opaque pagination cursor token returned by previous page request.
      example: "eyJpZCI6MTIzLCJjcmVhdGVkX2F0IjoxNjk4Nzc2MDAwfQ=="
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Maximum number of items to return per page (1-100, default 20).
      example: 20
    UserIdParam:
      name: user_id
      in: path
      required: true
      description: Identifier of the account the operation applies to.
      schema:
        type: string
        format: uuid
...
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      schema:
        type: string
      description: ETag header value for conditional HTTP request caching. Returns 304 Not Modified if matching.
      example: '"w/33a64df551425fcc55e4d42a148795d9f25f89d4"'
...
  headers:
    ETag:
      description: Entity tag representing the current version digest of the requested resource.
      schema:
        type: string
      example: '"w/33a64df551425fcc55e4d42a148795d9f25f89d4"'
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Standard machine-readable error code string identifier.
          example: "INVALID_INPUT"
        message:
          type: string
          description: Human-readable explanation of the error.
          example: "Email address is invalid or missing"
        details:
          type: object
          description: Structured key-value object containing detailed validation or field errors.
          additionalProperties: true
          example:
            field: "email"
            issue: "Must be a valid email format"
    CursorPagination:
      type: object
      required:
        - next_cursor
        - has_more
      properties:
        next_cursor:
          type: string
          nullable: true
          description: Opaque pagination token for requesting the next page. Null if no remaining items.
          example: "eyJpZCI6MTI0LCJjcmVhdGVkX2F0IjoxNjk4Nzc2MTAwfQ=="
        has_more:
          type: boolean
          description: Boolean flag indicating if additional records exist past this page.
          example: true
...
    UserEnvelope:
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
    UserListResponse:
      type: object
      required:
        - items
        - pagination
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        pagination:
          $ref: '#/components/schemas/CursorPagination'
...
    UserResponse:
      type: object
      required:
        - id
        - email
        - full_name
        - role
        - status
      properties:
        id:
          type: string
          format: uuid
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        email:
          type: string
          format: email
          example: "user@example.com"
        full_name:
          type: string
          example: "Jane Doe"
        role:
          type: string
          enum: [administrador, profesor, estudiante]
          example: "estudiante"
        status:
          type: string
          enum: [pending_verification, active, suspended]
          example: "active"
...
  responses:
    304NotModified:
      description: Resource has not changed since the last request (If-None-Match match).
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
    400BadRequest:
      description: Invalid request parameters or body payload syntax.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    401Unauthorized:
      description: Authentication required or invalid/expired session token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    403Forbidden:
      description: Insufficient permissions or role privileges for requested operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    404NotFound:
      description: The requested resource identifier was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    500InternalServerError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...